Verrrry basic Enemy AI.
Demo scene called EnemyTestScene has an example of the melee signless. Right now he runs at the player on sight, takes a bit of cooldown time, then charges at you if you are still in sight.
Also changed a bit of the motion logic so as to account for attacking in general, as well as the fact that some entities can be moved (as opposed to being walls that would always stop you). The latter was to prevent the player from getting stuck between a wall and an enemy.